+++ Src/Public/Invoke-AsBuiltReport.Veeam.VBR.ps1
@@ -423,6 +423,22 @@ function Invoke-AsBuiltReport.Veeam.VBR {
                     }
                 }
             }
+
+            #---------------------------------------------------------------------------------------------#
+            #                             Backup Jobs Restore Points Section                              #
+            #---------------------------------------------------------------------------------------------#
+            if ($InfoLevel.Jobs.PSObject.Properties.Value -ne 0) {
+                if (((Get-VBRJob -WarningAction SilentlyContinue).count -gt 0) -or ((Get-VBRTapeJob).count -gt 0) -or ((Get-VBRSureBackupJob).count -gt 0)) {
+                    Section -Style Heading2 'Restores Points' {
+                        Paragraph "The following section provides information about the jobs restore points in Veeam Server: $(((Get-VBRServerSession).Server))."
+                        BlankLine
+                        if ($InfoLevel.Jobs.Restores -gt 0) {
+                            Get-AbrVbrBackupJobsRP
+                            Get-AbrVbrTapeBackupJobsRP
+                        }
+                    }
+                }
+            }
         }
         if ((Get-VBRServerSession).Server) {
             Write-PScriboMessage "Disconecting section from $((Get-VBRServerSession).Server)"